backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.18
openai==1.3.7
pydantic==2.5.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
python-multipart==0.0.6
requests==2.31.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 